internal/database/postgres.go
package database

import (
	"fmt"

	"github.com/HermanPlay/backend/internal/config"
	"gorm.io/driver/postgres"
	"gorm.io/gorm"
)

type postgresDatabase struct {
	Db *gorm.DB
}

func NewPostgresDatabase(cfg *config.Config) (Database, error) {
	dsn := fmt.Sprintf(
		"postgres://%s:%s@%s:%v/%s?sslmode=disable",
		cfg.Db.User,
		cfg.Db.Password,
		cfg.Db.Host,
		cfg.Db.Port,
		cfg.Db.DBName,
	)
	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{})
	if err != nil {
		return nil, fmt.Errorf("cannot open db connection")
	}

	return &postgresDatabase{Db: db}, nil
}

func (p *postgresDatabase) Connect() *gorm.DB {
	return p.Db
}
